backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database - PostgreSQL with async support
asyncpg==0.29.0
SQLAlchemy==2.0.25
alembic==1.13.1

# Vector Database - Qdrant
qdrant-client==1.7.3

# Google Gemini API (for embeddings and generation)
google-generativeai==0.3.2

# Language Detection
langdetect==1.0.9

# Structured Logging
structlog==24.1.0

# Configuration Management
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.1

# HTTP Client
httpx==0.26.0

# Rate Limiting
slowapi==0.1.9

# Utilities
tenacity==8.2.3  # Circuit breakers and retries
python-dateutil==2.8.2

# Development Dependencies
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0
